Carry I2S audio as signed 16-bit end to end
The ring was 8-bit unsigned on both paths, so the APU mix was squeezed
through a byte and expanded again for a 16-bit DAC -- the quantisation
ceiling on the I2S path, audible as noise under quiet passages.

audio_sample_t follows the sink: still 8-bit unsigned with 128 silence for
PWM, which is what its DMA feeds the slice compare register, and signed
16-bit for I2S. The ring, the scratch buffer, the fill callback and
i2s_fill() all take that type, so the conversion at the end is now just a
duplicate into the two channels.

The mix keeps the channel sum at full width (0..1275) instead of averaging
it back down to 0..255, and maps full scale to 32767 at gain 100 -- the
same loudness at a given I2S_GAIN_PERCENT as before, with the
quantisation step 256x smaller. DC removal and saturation are unchanged.

Costs 10 KB of BSS (8 KB ring, the rest scratch): 194 KB of the RP2350's
520 KB. PWM targets are byte-for-byte as before; built WAVESHARE_LCD13 to
confirm.

+++ software/infones/i2s_output.h
@@ -11,9 +11,8 @@
  * Independent of the display: its own PIO state machine and DMA channels, so
  * it never touches the LCD's SPI or DMA channel.
  *
- * Samples are 8-bit unsigned mono, 128 = silence — the same format the PWM
- * path and InfoNES' audioRing already use — converted to signed 16-bit stereo
- * (both channels the same) on the way into the I2S FIFO.
+ * Samples are signed 16-bit mono, the same as InfoNES' audio ring carries on
+ * this path, duplicated into both channels on the way into the I2S FIFO.
  */
 
 #include <stdint.h>
@@ -22,9 +21,9 @@
 extern "C" {
 #endif
 
-/* Fill up to `count` mono 8-bit unsigned samples, return how many were
- * written. The remainder is padded with silence by the caller. */
-typedef int (*i2s_fill_fn)(uint8_t *dst, int count);
+/* Fill up to `count` mono signed 16-bit samples, return how many were
+ * written. Any shortfall is handled by the caller. */
+typedef int (*i2s_fill_fn)(int16_t *dst, int count);
 
 /* Claims the PIO SM and DMA channels on the first call (later calls only
  * restart the stream) and starts playing. */
